# server/api/openapi.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/connect": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Connect",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/rooms/{roomId}/join": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_JoinRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/messages": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_SendMessage",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/poke": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Poke",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PokeRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{roomId}/leave": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_LeaveRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConnectionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "connectionId": {
            "type": "string"
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "content": {
            "type": "string"
          },
          "from": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PokeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "targetConnectionId": {
            "type": "string"
          }
        }
      }
    }
  }
}